refactor: export strategies directly
BREAKING CHANGE: New manner of specifying a strategy.

Before:
```js
import PriorityQueue from 'priorityqueue';

const binaryHeap = new PriorityQueue({
  strategy: PriorityQueue.BinaryHeapStrategy
});
```

After:
```js
import PriorityQueue, { BinaryHeap } from 'priorityqueue';

const binaryHeap = new BinaryHeap();
PriorityQueue === BinaryHeap; // true
```

`PairingHeap` , `SkewHeap` also available.
