--- test/.eslintrc.js
@@ -1,13 +1,0 @@
-module.exports = {
-  "globals": {
-    "describe": false,
-    "it": false,
-    "test": false,
-    "beforeEach": false,
-    "afterEach": false,
-    "beforeAll": false,
-    "afterAll": false,
-    "expect": false,
-    "jest": false,
-  }
-};
